feat(cpu): attach cpu state to doc instance
Start the work on redesign doc metrics to carry state.
Attach cpu metric on doc instance.
Change the `data` event to `sample`.
Don't pass arguments to the event.

BREAKING CHANGE: the cpu metric is not exposed anymore to the
event handler, but it is attached to the Doc instance. The event name is
changed from `data` to `sample`.